add url

+++ controllers/wallabagController.php
@@ -84,7 +84,7 @@ class FreshExtension_wallabag_Controller extends Minz_ActionController {
 	
 	public function shareAction() {
 		$uri = $this->getEntriesUri();
-		$entryId = (int) Minz_Request::param('id');
+		$entryId = Minz_Request::param('id');
 		$entryDAO = FreshRSS_Factory::createEntryDao();
 		$entry = $entryDAO->searchById($entryId);
